+++ src/stripe/Components/InfoPriceColumn.js
@@ -19,6 +19,15 @@ const InfoPriceColumn = (props) => {
 		expirationDate,
 		discountedPrice,
 	} = props;
+	let messagePrice = ` ${__('per month', 'wubtitle')}`;
+	if (discountedPrice && discountedPrice.duration === 'once') {
+		messagePrice = ` ${__('for this month', 'wubtitle')}`;
+	}
+	if (discountedPrice && discountedPrice.duration === 'repeating') {
+		messagePrice = ` ${__('for the first', 'wubtitle')} ${
+			discountedPrice.durationInMonths
+		} ${__('months', 'wubtitle')}`;
+	}
 	let total = parseFloat(price);
 	if (taxable) {
 		total = parseFloat(price) + parseFloat(taxAmount);
@@ -44,9 +53,7 @@ const InfoPriceColumn = (props) => {
 						) : (
 							<span className="total">{total} &euro; </span>
 						)}
-						<span className="valxm">
-							{__('per month', 'wubtitle')}
-						</span>
+						<span className="valxm">{messagePrice}</span>
 						<InfoIcon
 							className="info-icon"
 							onClick={() => setIsOpen(!isOpen)}
@@ -71,6 +78,7 @@ const InfoPriceColumn = (props) => {
 						taxable={taxable}
 						total={total}
 						discountedPrice={discountedPrice}
+						messagePrice={messagePrice}
 					/>
 				)}
 			</div>
@@ -99,6 +107,7 @@ const InfoPriceColumn = (props) => {
 						taxable={taxable}
 						total={total}
 						discountedPrice={discountedPrice}
+						messagePrice={messagePrice}
 					/>
 				</div>
 				<Disclaimer />
